backend/requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg[binary]==3.2.1
gunicorn==22.0.0
celery==5.4.0
redis==5.0.7
whitenoise==6.7.0
pycodestyle==2.12.1
radon==6.0.1
cryptography==42.0.7
boto3==1.34.162
azure-identity==1.17.1
azure-mgmt-network==26.0.0
azure-mgmt-storage==22.1.0
azure-mgmt-keyvault==10.3.0
google-cloud-storage==2.18.2
google-cloud-compute==1.20.1
dnspython==2.6.1
